Perfil de las organizaciones de la BD finalizado (exceptos datos no definidos en la BD)

+++ app/src/main/java/com/example/proyecto_en_la_sombra/DAO/ClienteDAO.kt
@@ -34,7 +34,7 @@ interface ClienteDAO {
 
     //Get a client by a given id
     @Query("SELECT * FROM Cliente WHERE idCliente = :id")
-    suspend fun getClientById(id: Int): Cliente
+    suspend fun getClientById(id: Long): Cliente
 
     @Query("SELECT * FROM Cliente WHERE email = :email")
     suspend fun getClienteByEmail(email : String) : Cliente
